requirements.txt
# Build test: forzar lectura de runtime.txt
Flask==2.3.3
twilio==8.10.0
psycopg2-binary==2.9.7
PyJWT==2.10.1
requests==2.32.5
blinker==1.9.0
Werkzeug==3.1.3
Jinja2==3.1.6
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1